Delivery/backend/requirements.txt
Django==5.1.4
djangorestframework==3.15.2
djangorestframework-simplejwt==5.4.0
django-cors-headers==4.6.0
drf-spectacular==0.28.0
Pillow==11.0.0
django-filter==24.3
channels==4.2.0
